requirements.txt
numpy
pandas
matplotlib
scikit-learn
torch
torchvision
torchaudio
pyyaml
tqdm
jupyter
